Gen_AI_projects/sagemaker_llm_ft/requirements.txt
jupyter
transformers
sagemaker
pandas
numpy
python-dotenv
scikit-learn
matplotlib
seaborn

